Rename to cswitch, add items in queue
